+++ src/main/java/dev/willyelton/crystal_tools/levelable/tool/BowLevelableItem.java
@@ -12,10 +12,13 @@ import net.minecraft.stats.Stats;
 import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.AbstractArrow;
+import net.minecraft.world.entity.projectile.Arrow;
 import net.minecraft.world.item.*;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
@@ -46,25 +49,25 @@ public class BowLevelableItem extends BowItem implements LevelableItem {
                 itemstack = getProjectile(stack, player);
             }
 
-            int i = this.getUseDuration(stack) - timeLeft;
-            i = net.minecraftforge.event.ForgeEventFactory.onArrowLoose(stack, level, player, i, !itemstack.isEmpty() || flag);
-            if (i < 0) return;
+            int timeUsed = this.getUseDuration(stack) - timeLeft;
+            timeUsed = net.minecraftforge.event.ForgeEventFactory.onArrowLoose(stack, level, player, timeUsed, !itemstack.isEmpty() || flag);
+            if (timeUsed < 0) return;
 
             if (!itemstack.isEmpty() || flag) {
                 if (itemstack.isEmpty()) {
                     itemstack = new ItemStack(Items.ARROW);
                 }
 
-                float f = getPower(i, stack);
-                if (((double)f >= 0.1D)) {
+                float power = getPower(timeUsed, stack);
+                if (((double) power >= 0.1D)) {
                     boolean flag1 = player.getAbilities().instabuild || (itemstack.getItem() instanceof ArrowItem && ((ArrowItem)itemstack.getItem()).isInfinite(itemstack, stack, player));
                     if (!level.isClientSide) {
-                        ArrowItem arrowitem = (ArrowItem)(itemstack.getItem() instanceof ArrowItem ? itemstack.getItem() : Items.ARROW);
+                        ArrowItem arrowitem = (ArrowItem) (itemstack.getItem() instanceof ArrowItem ? itemstack.getItem() : Items.ARROW);
                         AbstractArrow abstractarrow = arrowitem.createArrow(level, itemstack, player);
                         abstractarrow = customArrow(abstractarrow);
                         //TODO: TOO Random
-                        abstractarrow.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, f * 3.0F + NBTUtils.getFloatOrAddKey(stack, "arrow_speed_bonus") / 4.0F, 0.0F);
-                        if (f == 1.0F) {
+                        abstractarrow.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, power * 3.0F + NBTUtils.getFloatOrAddKey(stack, "arrow_speed_bonus") / 4.0F, 0.0F);
+                        if (power == 1.0F) {
                             abstractarrow.setCritArrow(true);
                         }
 
@@ -82,6 +85,18 @@ public class BowLevelableItem extends BowItem implements LevelableItem {
                             abstractarrow.setSecondsOnFire(100);
                         }
 
+                        if (abstractarrow instanceof Arrow arrow && stack.getTag() != null) {
+                            for (int i = 1; i < 34; i++) {
+                                if (stack.getTag().contains("effect_" + i)) {
+                                    MobEffect effect = MobEffect.byId(i);
+                                    if (effect != null) {
+                                        MobEffectInstance instance = new MobEffectInstance(effect, (int) NBTUtils.getFloatOrAddKey(stack, "effect_" + i) * 20, 1, false, true);
+                                        arrow.addEffect(instance);
+                                    }
+                                }
+                            }
+                        }
+
                         stack.hurtAndBreak(1, player, (p_40665_) -> p_40665_.broadcastBreakEvent(player.getUsedItemHand()));
                         if (flag1 || player.getAbilities().instabuild && (itemstack.is(Items.SPECTRAL_ARROW) || itemstack.is(Items.TIPPED_ARROW)) || NBTUtils.getFloatOrAddKey(stack, "infinity") > 0) {
                             abstractarrow.pickup = AbstractArrow.Pickup.CREATIVE_ONLY;
@@ -90,7 +105,7 @@ public class BowLevelableItem extends BowItem implements LevelableItem {
                         level.addFreshEntity(abstractarrow);
                     }
 
-                    level.playSound(null, player.getX(), player.getY(), player.getZ(), SoundEvents.ARROW_SHOOT, SoundSource.PLAYERS, 1.0F, 1.0F / (level.getRandom().nextFloat() * 0.4F + 1.2F) + f * 0.5F);
+                    level.playSound(null, player.getX(), player.getY(), player.getZ(), SoundEvents.ARROW_SHOOT, SoundSource.PLAYERS, 1.0F, 1.0F / (level.getRandom().nextFloat() * 0.4F + 1.2F) + power * 0.5F);
                     if (!flag1 && !player.getAbilities().instabuild) {
                         itemstack.shrink(1);
                         if (itemstack.isEmpty()) {